- fix signed-unsigned comparisor

+++ src/test/lib/BasicTest.cpp
@@ -284,7 +284,7 @@ void QuicTestAddrFunctions(_In_ int Family)
         TEST_TRUE((SockAddr.Ipv4.sin_addr.s_addr & 0x00FFFF00UL) == 0);
     } else {
         memset(&SockAddr.Ipv6.sin6_addr, 0XFF, sizeof(SockAddr.Ipv6.sin6_addr));
-        for (int i = 0; i < sizeof(SockAddr.Ipv6.sin6_addr.s6_addr) - 1; i++) {
+        for (unsigned long i = 0; i < sizeof(SockAddr.Ipv6.sin6_addr.s6_addr) - 1; i++) {
             TEST_TRUE(SockAddr.Ipv6.sin6_addr.s6_addr[i] == 0);
         }
     }
